+++ hydra-node/src/Hydra/Chain/Direct/Tx.hs
@@ -118,6 +118,34 @@ instance Arbitrary ClosedThreadOutput where
 hydraHeadV1AssetName :: AssetName
 hydraHeadV1AssetName = AssetName (fromBuiltin hydraHeadV1)
 
+-- | The metadata label used for identifying Hydra protocol transactions. As
+-- suggested by a friendly large language model: The number most commonly
+-- associated with "Hydra" is 5, as in the mythological creature Hydra, which
+-- had multiple heads, and the number 5 often symbolizes multiplicity or
+-- diversity. However, there is no specific numerical association for Hydra
+-- smaller than 10000 beyond this mythological reference.
+hydraMetadataLabel :: Word64
+hydraMetadataLabel = 55555
+
+-- | Create a transaction metadata entry to identify Hydra transactions (for
+-- informational purposes).
+mkHydraHeadV1TxName :: Text -> TxMetadata
+mkHydraHeadV1TxName name =
+  TxMetadata $ Map.fromList [(hydraMetadataLabel, TxMetaText $ "HydraV1/" <> name)]
+
+-- | Get the metadata entry to identify Hydra transactions (for informational
+-- purposes).
+getHydraHeadV1TxName :: Tx -> Maybe Text
+getHydraHeadV1TxName =
+  lookupName . txMetadata . getTxBodyContent . getTxBody
+ where
+  lookupName = \case
+    TxMetadataNone -> Nothing
+    TxMetadataInEra (TxMetadata m) ->
+      case Map.lookup hydraMetadataLabel m of
+        Just (TxMetaText name) -> Just name
+        _ -> Nothing
+
 -- * Create Hydra Head transactions
 
 -- | Create the init transaction from some 'HeadParameters' and a single TxIn
@@ -139,6 +167,7 @@ initTx networkId seedTxIn participants parameters =
             : map (mkInitialOutput networkId seedTxIn) participants
         )
       & mintTokens (HeadTokens.mkHeadTokenScript seedTxIn) Mint ((hydraHeadV1AssetName, 1) : participationTokens)
+      & setTxMetadata (TxMetadataInEra $ mkHydraHeadV1TxName "InitTx")
  where
   participationTokens =
     [(onChainIdToAssetName oid, 1) | oid <- participants]
@@ -202,6 +231,7 @@ commitTx networkId scriptRegistry headId party utxoToCommitWitnessed (initialInp
       & addInputs committedTxIns
       & addExtraRequiredSigners [vkh]
       & addOutputs [commitOutput]
+      & setTxMetadata (TxMetadataInEra $ mkHydraHeadV1TxName "CommitTx")
  where
   initialWitness =
     BuildTxWith $
@@ -273,6 +303,7 @@ collectComTx networkId scriptRegistry vk headId headParameters (headInput, initi
       & addReferenceInputs [commitScriptRef, headScriptRef]
       & addOutputs [headOutput]
       & addExtraRequiredSigners [verificationKeyHash vk]
+      & setTxMetadata (TxMetadataInEra $ mkHydraHeadV1TxName "CollectComTx")
  where
   HeadParameters{parties, contestationPeriod} = headParameters
 
@@ -360,6 +391,7 @@ closeTx scriptRegistry vk closing startSlotNo (endSlotNo, utcTime) openThreadOut
       & addExtraRequiredSigners [verificationKeyHash vk]
       & setValidityLowerBound startSlotNo
       & setValidityUpperBound endSlotNo
+      & setTxMetadata (TxMetadataInEra $ mkHydraHeadV1TxName "CloseTx")
  where
   OpenThreadOutput
     { openThreadUTxO = (headInput, headOutputBefore)
@@ -452,6 +484,7 @@ contestTx scriptRegistry vk Snapshot{number, utxo} sig (slotNo, _) closedThreadO
       & addOutputs [headOutputAfter]
       & addExtraRequiredSigners [verificationKeyHash vk]
       & setValidityUpperBound slotNo
+      & setTxMetadata (TxMetadataInEra $ mkHydraHeadV1TxName "ContestTx")
  where
   ClosedThreadOutput
     { closedThreadUTxO = (headInput, headOutputBefore)
@@ -528,6 +561,7 @@ fanoutTx scriptRegistry utxo (headInput, headOutput) deadlineSlotNo headTokenScr
       & addOutputs orderedTxOutsToFanout
       & burnTokens headTokenScript Burn headTokens
       & setValidityLowerBound (deadlineSlotNo + 1)
+      & setTxMetadata (TxMetadataInEra $ mkHydraHeadV1TxName "FanoutTx")
  where
   headWitness =
     BuildTxWith $